Move compiler detection related code to a new CompilerChoice class
